# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quest on Agora API
  version: 0.10.2
  description: |
    REST API for discussion sessions, participants, AI chat, billing, and admin.
    Canonical discussion endpoints use `/api/discussions` (plural). Legacy `/api/discussion` routes are deprecated.
servers:
  - url: /api
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
...
paths:
...
  /checkout/toss/callback:
    post:
      tags:
        - Billing
      summary: Toss payment callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Processed
          content:
            application/json:
              schema:
                type: object
...